Allow a project to contain multiple repositories
Add a Repository type to the database so that a project can have
more than one repository. Rework the main program and the Hammer
interface to support this.

All the data from all the repositories will be stored in the
database, but the commit iteration has not been updated; it picks
an arbitrary repository to iterate over. This will be fixed soon.