Add id of head commit to repository
When only iterating over commits, there should be no need to
refresh the database. The old way of accessing the current head
of the actual git repository required an update so that the head
was also present in the database. Switch to storing the last
known head commit id in the repository so that iteration can
start from that and proceed with only the data that has already
been written to the database.