Process commits from earliest to latest
Starting the commit processing from the head commit and going
backwards caused problems with author attribution in certain
merge commits. So switch to processing commits from the earliest
one, process them in an order where all parents are processed
before any children, and run the diffed commit stats of merge
commits against the main branch. This gives the same results
as running a full blame, it doesn't produce negative line counts,
and is a bit faster to boot.