Implement periodic database commits
When running on a very large repository, it's a good idea to
periodically commit the computed data into the database so that
the run can be interrupted and continued later, or that bugs in
git-hammer don't waste all the work that was already done. So
add code to commit the session to the database every 5 minutes.
Since the existing commits are already ignored on updates, this
is enough to support restarting later.