Fix timezones for hour and weekday graphs
It does not seem that Postgres actually stores the original
timezone of the timestamp, even when timezone storing is on.
Since having the correct timezone is crucial to hour-of-day and
day-of-week determination, store the timezone of a commit also
in the database, and use this timezone to create correctly-zoned
datetimes when generating these graphs.

I'm not sure which other places should switch to using the new
timezone-aware commit time of a commit. Maybe all of them? Maybe
this is enough? I'm being cautious and not modifying anything
else quite yet, but this will need to be checked if there ever
is a more precise checking mechanism.