Implement basic tests for Hammer object initialization
Change the Hammer object implementation so that only creating
a Hammer object does not yet create the database if it did not
already exist. Move database creation instead to the
add_repository method, and add verifications to other methods
that the database has been created. Test these new features with
unit tests.