Add script to verify regressions against a known-good database
When trying to optimize performance, it is really important to
not break the calculations that are being performed. To make this
easier to verify, add a script that takes two project names for
the same repository (intended to be a known-good version and a
new improved version) and verifies that the data computed for
each commit is the same.

Future improvements: Verify the whole database, though verifying
the commits does the important parts. Run this automatically
(will need to wait until proper CI is set up anyway).