Produce proper commit objects when iterating over all commits
The iter_individual_commits method that allowed iterating over
all the commits stored in the database was producing the commits
as they are stored in the database. This meant that the line_counts
and test_counts attributes weren't populated, which reduced the
utility of the regression checking.

Instead of database-level commits, produce the processed commit
objects that have the line and test counts computed. This added
two new requirements:
1) Implement __eq__ and __hash__ for authors so that the
comparison in regression checking works for the count
dictionaries.
2) Load the Author relationship of Commit eagerly so that the
author of a commit is available afterwards.