Ignore submodules when diffing
By default, git diff includes entries for submodules in its output.
This breaks when trying to extract information from such a file
since the submodule directory is not in fact a file known to git.
This may be related to a GitPython issue:
https://github.com/gitpython-developers/GitPython/issues/891

But for git-hammer, the sensible course of action seems to be
ignoring submodules. They are not code in the main repository,
and if their code should be in a project, the actual repository
can be included. So add to diff the option ignore-submodules that
causes it to skip submodules.

Also add a unit test that creates a submodule in the test
repository and verifies that git-hammer works correctly on it.