Ignore decoding errors when reading source lines
To be able to match a code line against the test line regex, it
needs to be decoded into a string. This decoding is done by
assuming UTF-8, which broke for files that were using some other
encoding.

Use the 'ignore' errors parameter for line decoding so that
files not encoded in UTF-8 can be processed. This should be fine
since the contents of the lines are only used to match against
the test line regex, which typically is in the ASCII range. This
will need to be revisited if it turns out to break in some cases.

Also added a test that the lines in the initial commit are
counted correctly.