Make sure test repositories are cloneable
A previous gc run on the unit test repository had removed all the
files from the refs directory, which to git means the directory
no longer exists. However, the presence of the refs directory is
required for git to be able to see it as a repository, so all the
tests that were cloning the repository were broken.

Fix the problem by committing an empty .keep file in the refs
directories of the test repositories. This seems to work, as git
doesn't seem to care about extra files in the repository, only
the ones that it works with.