Make summary functions return results instead of displaying
To make Git Hammer function properly as a library, its summary
functions should not be trying to display the results. Instead,
they should return a suitable object and let the caller decide
what to do with it.

Change the graph functions to return the Matplotlib Figure object
instead of showing it. Change the table functions to return the
beautifultable object instead of printing it. And adjust the
main program to keep functioning in the same manner.