Add naming convention to database constraints and indexes
To be able to perform migrations involving constraints or
indexes, it is necessary to know the name of the constraint or
index. By default, SQLAlchemy does not assign any name, leaving
it to the database. Since each database has its own internal
naming convention for these, it is not possible to write correct
migrations.

Create a MetaData object for the SQLAlchemy classes containing
a naming convention so that any constraints and indexes will
receive predictable names and can thus be used in migrations.