Allow one database to contain multiple projects
In some potential use cases, such as running on some cloud
services, it is not feasible to create a new database for each
new project. This means that a single database should be able to
contain multiple projects.

Add a Project table to the database. Link repositories to projects
through an additional table, since a single repository might be
included in multiple projects. Keep the Hammer object handling
only a single project (with the exception of a new function for
querying all project names).

A minor change, that unfortunately made it impossible to create
a migration, was to change the primary key of the repository
table to an autoincrementing integer.

New tests were added to test multi-project functionality, and
the regression checking script updated to require the checked
projects to be in different databases (since the checked projects
have the same repositories, it makes no sense for regression
checking if they were in the same database).