Update documentation for multi-project support
Update the README slightly because now the URL to the Hammer
object is to the database, not to the server. Also update the
main program to use the appropriate environment variable.

Add a note to the top of the README to inform users that the
database schema changed in a way that requires rebuilding any
existing databases.