Refactor repository configuration into its own object
The repository configuration is treated as a plain dictionary,
it is used by some support functions, and its creation and use is
distributed across a few places. It makes sense to refactor it
into its own object that knows how to construct it from JSON, and
how to use its properties to recognize source files and tests.