Add function and command to iterate over sources and tests
It is difficult to know exactly what files are counted as source
or test files, and what lines counted as test lines. To improve
this situation, implement a function that iterates over all the
files in the repository, returning the ones that are designated
source or test files by the configuration, and furthermore,
iterate over the lines in each test file, returning those lines
that are designated as test lines. Also make this into a command
for the main program.