Support shallow clones
When run on a shallow clone, that is, one where not all commits
are cloned into the repository, Git Hammer would crash. This was
because the SHAs of the excluded commits would still show up as
parents of the included commits, and Git Hammer would try to use
those SHAs in its git commands.

Fix this by checking whether a parent commit is actually present
before trying to run diffs against it. This will allow Git
Hammer to process a shallow clone, but naturally the statistics
will be based on what data is available, and they will not be
updated if the repository is unshallowed later. But updating
existing data is not yet something that Git Hammer does in any
situation.