Add tests for combined commits
Create test cases where multiple repositories are added to one
project, and check that the combined commits produced from this
project are created with the correct attributes.

This revealed an odd threading issue, probably with Author
database objects. Switch to using load=True for these objects,
since that appears to fix the problem, even though it is slower.