Exclude non-unit-tested files from coverage report
The main program for githammer and the summary functions are
simple code that is typically only executed in a UI context.
Therefore, it doesn't seem important to ever write unit tests
for them, so they can be excluded from the test coverage report.

If there is ever any need to unit-test some parts of them, it's
likely a better idea to extract the unit-testable part into
githammer proper.