Add ability to limit commit processing by date
Add an optional parameter to add_repository that allows excluding
commits older than the specified date from processing. This is
helpful for large repositories where processing the full commit
history would take a long time and is not needed.

The exclusion date is best stored in the database, so that future
updates don't try to unexpectedly process the originally-excluded
commits. Therefore there is also an alembic migration for the
database.

The functionality is not yet complete. The only test is that the
processed commits are limited to those that are not excluded.
Further work will make sure that the included commits are
processed correctly.