Run full blame on commits with parent in the repository
When limiting commits that are processed at the Git Hammer level,
the initial processed commits will still have parents in the
repository, which means that git can construct proper attributions
for the lines. In this case, assigning all the lines in the
initial commit to its author is not necessary, as better
information is available.

Check, when processing a parentless commit, whether its parent
does exist in the repository. In this case, run a full git blame
to assign the lines as accurately as possible from the repository
information instead of defaulting the assignment to the commit
author. Also normalize the line and test counts in this case to
avoid storing 0 values.