Limit number of authors listed in per-author graph
The legend of the per-author graph is not very useful when there
are many authors. To make it somewhat more useful, display only
the authors that are in the top 25 by count in the head commit.
This also changes the ordering in the graph to be by head commit
count instead of by author name, which incidentally makes the
graph easier to read too.