Allow one repository to be in multiple projects
Since commits are identified by their SHA, one commit can only be
added to the database once. But if multiple projects include the
same repository (for instance, a library that is shared), that
repository needs to be a part of all projects.

Fix the situation by checking whether a repository to be added
to a project already exists in the database, and only add the
commits instead of trying to process again the commits in the
database. Verify in a test that the repository is indeed added
to all the projects.