Fix submodule handling in the initial commit
When handling the initial commit, Git Hammer uses the traverse
function of the tree object, passing it the visit_once parameter
to make sure each file is processed only once. Unfortunately this
breaks when the tree contains submodules, since the visited check
in traverse raises an exception because of a GitPython issue
(https://github.com/gitpython-developers/GitPython/issues/597).

Remove the visit_once parameter from the traverse call. Since
it's the default for a tree object, it should be safe. Also add
pruning of submodules and a test to verify that submodules in the
initial commit are correctly ignored.